- move canonical temporary Gemini execution to stateless chat - deprecate the legacy temporary chat endpoint - support valid slash-containing Gemini WebAPI model IDs - share direct WebAPI model classification across validation and discovery - preserve temporary endpoint compatibility - update API documentation and tests
