+++ pyproject.toml
@@ -52,7 +52,7 @@ asyncio_default_test_loop_scope = "function"
 timeout = 30
 # Temporary workaround for Starlette using deprecated anyio.abc.BlockingPortal.
 # Remove once Starlette no longer emits this warning.
-Tracking: #142
+# Tracking: #142
 filterwarnings = [
     "ignore:The anyio\\.abc\\.BlockingPortal alias is deprecated, use anyio\\.from_thread\\.BlockingPortal instead\\.:DeprecationWarning:starlette\\.testclient",
 ]
