+++ src/app/endpoints/chat.py
@@ -20,7 +20,10 @@ from app.services.factory import ProviderFactory
 from app.services.model_catalog import list_models as build_model_catalog
 from app.services.model_catalog import list_stateless_models as build_stateless_model_catalog
 from app.services.openai_compatibility import validate_openai_request_compatibility
-from app.services.providers.gemini.temporary_chat import handle_temporary_chat_completions
+from app.services.providers.gemini.stateless_chat import handle_stateless_chat_completions
+from app.services.providers.gemini.temporary_chat import (
+    handle_temporary_chat_completions,
+)
 from app.services.providers.gemini.shared import (
     ensure_gemini_client_ready,
     validate_direct_webapi_model_name,
@@ -93,15 +96,18 @@ async def translate_chat(request: GeminiRequest):
 @router.post(
     "/v1/temporary/chat/completions",
     tags=["Chat"],
-    summary="Temporary OpenAI-Compatible Chat Completions",
+    summary="Temporary OpenAI-Compatible Chat Completions (Deprecated)",
+    deprecated=True,
     description=(
-        "Gemini WebAPI-only OpenAI-compatible chat completions endpoint. Requests are sent with temporary=True, "
-        "so responses are not saved in Gemini history and do not write SQLite conversation snapshots. "
-        "`conversation_id` is rejected. Playwright models/providers, Atlas models/providers, and any non-Gemini provider are rejected. "
-        "Malformed audited OpenAI controls return HTTP 422; controls unsupported by Gemini WebAPI return HTTP 400. "
-        "The endpoint supports streaming and non-streaming responses. File content parts are supported only by "
-        "Gemini WebAPI, are request-scoped, and generated artifact metadata follows the same response shape as "
-        "`/v1/chat/completions`."
+        "Deprecated compatibility endpoint. New integrations must use the canonical "
+        "`/v1/stateless/chat/completions` endpoint. This endpoint remains for backward "
+        "compatibility and delegates to the same stateless Gemini WebAPI implementation "
+        "(temporary=True, so responses are not saved in Gemini history and do not write "
+        "SQLite conversation snapshots; client-owned history, `conversation_id` is rejected). "
+        "Gemini WebAPI-only; Playwright, Atlas, and non-Gemini providers are rejected. "
+        "Malformed audited OpenAI controls return HTTP 422; controls unsupported by Gemini "
+        "WebAPI return HTTP 400. streaming and non-streaming responses, file parts, and "
+        "generated artifact metadata follow the same response shape as `/v1/chat/completions`."
     ),
     responses={
         200: CHAT_COMPLETIONS_RESPONSE_200,
@@ -126,15 +132,17 @@ async def temporary_chat_completions(request: OpenAIChatRequest):
     tags=["Chat"],
     summary="Stateless OpenAI-Compatible Chat Completions",
     description=(
-        "Client-owned-history chat completions endpoint for direct Gemini WebAPI execution only. Every request is "
+        "Canonical stateless Gemini WebAPI endpoint. Client-owned-history execution: every request is "
         "self-contained, uses temporary=True, rejects `conversation_id`, does not create SQLite conversation "
-        "snapshots, and does not persist Gemini conversation history. Playwright, Atlas, and other non-Gemini "
-        "providers are not supported. Malformed request values return HTTP 422; unsupported Gemini controls, "
-        "providers, backends, and `provider_options.gemini` return HTTP 400. Buffered responses, progressive SSE "
-        "streaming, multimodal file parts, and one generated function tool call are supported. `stream=true` with "
-        "tools uses buffered OpenAI-compatible SSE replay rather than native progressive tool streaming. "
-        "`max_tokens`, `max_completion_tokens`, `reasoning_effort`, and `stream_options.include_usage` are accepted "
-        "compatibility no-ops. Direct Gemini WebAPI execution has a 300-second deadline."
+        "snapshots, and does not persist Gemini conversation history. Gemini WebAPI only; Playwright, Atlas, "
+        "and other non-Gemini providers are not supported. Slash-containing model IDs are valid when advertised "
+        "by `/v1/stateless/models` and recognized as available by the Gemini WebAPI runtime catalog; unknown "
+        "slash IDs and non-Gemini routing IDs are rejected. Malformed request values return HTTP 422; unsupported "
+        "Gemini controls, providers, backends, and `provider_options.gemini` return HTTP 400. Buffered responses, "
+        "progressive SSE streaming, multimodal file parts, and one generated function tool call are supported. "
+        "`stream=true` with tools uses buffered OpenAI-compatible SSE replay rather than native progressive tool "
+        "streaming. `max_tokens`, `max_completion_tokens`, `reasoning_effort`, and `stream_options.include_usage` "
+        "are accepted compatibility no-ops. Direct Gemini WebAPI execution has a 300-second deadline."
     ),
     responses={
         200: CHAT_COMPLETIONS_RESPONSE_200,
@@ -152,7 +160,7 @@ async def temporary_chat_completions(request: OpenAIChatRequest):
     },
 )
 async def stateless_chat_completions(request: OpenAIChatRequest):
-    return await handle_temporary_chat_completions(
+    return await handle_stateless_chat_completions(
         request,
         endpoint_name="stateless",
         direct_webapi_only=True,
@@ -164,7 +172,8 @@ async def stateless_chat_completions(request: OpenAIChatRequest):
     tags=["Chat"],
     summary="List Stateless Models",
     description=(
-        "Returns only currently available direct Gemini WebAPI models that satisfy the stateless execution contract. "
+        "Returns only currently available direct Gemini WebAPI models that satisfy the stateless execution contract, "
+        "including valid slash-containing model IDs when advertised by the Gemini WebAPI runtime catalog. "
         "Playwright stateless execution is not implemented; Playwright models, legacy browser aliases, Atlas models, "
         "and other provider models are not included."
     ),
